Port GeoItems to Common using RenderProvider and two mixins
