+++ fabric/src/main/java/software/bernie/example/client/renderer/armor/GeckoArmorRenderer.java
@@ -1,9 +1,14 @@
 package software.bernie.example.client.renderer.armor;
 
+import com.mojang.blaze3d.vertex.PoseStack;
+import com.mojang.blaze3d.vertex.VertexConsumer;
+import net.minecraft.client.renderer.MultiBufferSource;
 import net.minecraft.resources.ResourceLocation;
+import org.joml.Vector3d;
 import software.bernie.example.item.GeckoArmorItem;
 import software.bernie.example.item.WolfArmorItem;
 import software.bernie.geckolib.GeckoLib;
+import software.bernie.geckolib.cache.object.BakedGeoModel;
 import software.bernie.geckolib.model.DefaultedItemGeoModel;
 import software.bernie.geckolib.renderer.GeoArmorRenderer;
 import software.bernie.geckolib.renderer.GeoRenderer;
