--- fabric/src/main/java/software/bernie/geckolib/mixins/client/ItemRendererAccessor.java
@@ -1,12 +1,0 @@
-package software.bernie.geckolib.mixins.client;
-
-import net.minecraft.client.renderer.BlockEntityWithoutLevelRenderer;
-import net.minecraft.client.renderer.entity.ItemRenderer;
-import org.spongepowered.asm.mixin.Mixin;
-import org.spongepowered.asm.mixin.gen.Accessor;
-
-@Mixin(ItemRenderer.class)
-public interface ItemRendererAccessor {
-    @Accessor("blockEntityRenderer")
-    BlockEntityWithoutLevelRenderer getBlockEntityRenderer();
-}
