+++ fabric/src/main/resources/geckolib.mixins.json
@@ -8,6 +8,8 @@
   "mixins": [
   ],
   "client": [
+    "client.MixinHumanoidArmorLayer",
+    "client.MixinItemRenderer",
     "client.TextureManagerMixin"
   ]
 }
