--- forge/src/main/java/software/bernie/example/entity/ReplacedCreeperEntity.java
@@ -1,37 +1,0 @@
-package software.bernie.example.entity;
-
-import net.minecraft.world.entity.EntityType;
-import software.bernie.geckolib.animatable.GeoEntity;
-import software.bernie.geckolib.animatable.GeoReplacedEntity;
-import software.bernie.geckolib.constant.DefaultAnimations;
-import software.bernie.geckolib.core.animatable.instance.AnimatableInstanceCache;
-import software.bernie.geckolib.core.animation.AnimatableManager;
-import software.bernie.geckolib.util.GeckoLibUtil;
-
-/**
- * Replacement {@link net.minecraft.world.entity.monster.Creeper} {@link GeoEntity} to showcase
- * replacing the model and animations of an existing entity
- * @see software.bernie.geckolib.renderer.GeoReplacedEntityRenderer
- * @see software.bernie.example.client.renderer.entity.ReplacedCreeperRenderer
- * @see software.bernie.example.client.model.entity.ReplacedCreeperModel
- */
-public class ReplacedCreeperEntity implements GeoReplacedEntity {
-	private final AnimatableInstanceCache cache = GeckoLibUtil.createInstanceCache(this);
-
-	// Register the idle + walk animations for the entity.<br>
-	// In this situation we're going to use a generic controller that is already built for us
-	@Override
-	public void registerControllers(AnimatableManager.ControllerRegistrar controllers) {
-		controllers.add(DefaultAnimations.genericWalkIdleController(this));
-	}
-
-	@Override
-	public AnimatableInstanceCache getAnimatableInstanceCache() {
-		return cache;
-	}
-
-	@Override
-	public EntityType<?> getReplacingEntityType() {
-		return EntityType.CREEPER;
-	}
-}
