--- neoforge/src/main/java/software/bernie/example/client/renderer/item/JackInTheBoxRenderer.java
@@ -1,16 +1,0 @@
-package software.bernie.example.client.renderer.item;
-
-import net.minecraft.resources.ResourceLocation;
-import software.bernie.example.item.JackInTheBoxItem;
-import software.bernie.geckolib.GeckoLib;
-import software.bernie.geckolib.model.DefaultedItemGeoModel;
-import software.bernie.geckolib.renderer.GeoItemRenderer;
-
-/**
- * Example {@link software.bernie.geckolib.renderer.GeoItemRenderer} for {@link JackInTheBoxItem}
- */
-public class JackInTheBoxRenderer extends GeoItemRenderer<JackInTheBoxItem> {
-	public JackInTheBoxRenderer() {
-		super(new DefaultedItemGeoModel<>(new ResourceLocation(GeckoLib.MOD_ID, "jack_in_the_box")));
-	}
-}
