+++ neoforge/src/main/resources/geckolib.mixins.json
@@ -9,6 +9,8 @@
   "mixins": [
   ],
   "client": [
-    "client.TextureManagerMixin"
+    "client.TextureManagerMixin",
+    "client.MixinHumanoidArmorLayer",
+    "client.MixinItemRenderer"
   ]
 }
